Bug 1465841: handle gecko-actions, gecko-roles features
This also moves the extra_tc_scopes into the `repo:..` role, rather than in the
trust-level-specific role.  I anticipate future work will actually get rid of
all this complex role inheritance, so this is probably a taste of the future.

It also grants active_scm_level_1 access to nss-try's scopes, and
active_scm_level_3 access to ci-admin/ci-configuration's scopes.  I don't think
this is an issue.