Bug 1526979 - add support for hg-push tasks r=tomprince
This adds support for "bindings" in the Hooks API, and uses it to support hooks
that run when a push is generated, with the hook template based on a file in
ci-configuration.

The hooks run tasks on a dedicated `hg-push` workerType which *only* runs this
sort of task, and can thus run them quickly.  The tasks themselves do not
perform an hg clone or do anything terribly complicated, so once the docker
image is loaded they only take a few seconds to execute.

The docker image is generated on every push to the ci-admin repository.  That
could be attached to the ci-configuration repository, instead, but that
repository does not yet use ci-taskgraph.

This also adds some checks to be confident that nobody can trigger hg-push
hooks, nor create tasks on the workerType.

Differential Revision: https://phabricator.services.mozilla.com/D21282