Bug 1573987 - Break generic bits of ci-admin out into tc-admin; r=tomprince
The entire project is divided into `tcadmin/`, which implements a generic
library for administering TC using the familiar patterns: generated, current,
diff, apply, check; and `ciadmin/`, which contains the implementation of
Firefox-CI administration using that library. There are no imports of
`ciadmin.*` from `tcadmin.*`.

The tc-admin library is designed to use a top-level `tc-admin.py` discoverd by
the `tc-admin` command, but to maintain compatibility with existing command
lines, the `ci-admin` command generates the appconfig directly and hands it to
the tcadmin library.

This is a set-up for subsequent work to make tc-admin a separate Python
distribution package, on which ci-admin depends.

Differential Revision: https://phabricator.services.mozilla.com/D44731