diff: support unknown starting characters when colouring diff
We switched to relying on the system diff late last year, but this also
means we shouldn't make assumptions about what characters the diff lines
can start with.

For example, sometimes the diff binary has a `\\ No newline at end of
file` in it (for some reason). When this happens an exception is raised
because `\\` is not a valid key in the 'colors' dictionary.

This turns 'colors' into a `defaultdict` so any lines that start with
unrecognized characters will simply be passed through uncoloured.

Fixes #204