diff: exit 2 in case we find differences
There's currently no way for consumers to differentiate between "there
was an error generating the diff" and "there were differences", since
both exit 1.

Unfortunately this is reversed from the diff command, which exits "0 if
the inputs are the same, 1 if different, 2 if trouble", but given
python's behavior I'm not sure we can reliably do better.